cm 28/07/2017 update/improve set_up.sh / option to keep log files / remove emails to oneself

+++ machine_code/execution_master.sh
@@ -44,7 +44,7 @@ cd ${wd_path_code}/stage_b
 
 ## classify
 R CMD BATCH --no-save "--args ${init_path} ${execution_id} ${data_path_train} \
-${data_path_raw_outbox} ${data_path_temp} ${wd_path_log} ${wd_path_model} ${lib_path}" classify.R \
+${data_path_raw_outbox} ${data_path_temp} ${wd_path_log} ${wd_path_model}" classify.R \
 ${wd_path_log}/classify_${execution_id}.Rout
 
 ## delete output file
@@ -60,7 +60,7 @@ cd ${wd_path_code}/stage_b
 
 ## aggregate
 R CMD BATCH --no-save "--args ${init_path} ${execution_id} ${data_path_temp} \
-${wd_path_helper_email} ${wd_path_log} ${lib_path}" aggregate.R ${wd_path_log}/aggregate_${execution_id}.Rout
+${wd_path_helper_email} ${wd_path_log} ${email_address}" aggregate.R ${wd_path_log}/aggregate_${execution_id}.Rout
 
 ## delete output file
 [ -e .RData ] && rm .RData
@@ -73,20 +73,29 @@ printf "\n# Clearing Data\n"
 cd ${data_path_raw_outbox}
 rm email*txt
 
+## create output folder
+mkdir ${wd_path_output}/output_${current_date}_${execution_id}
+
 ## move output to location & delete all data
 cd ${data_path_temp}
-mv *mean_plot_${execution_id}* $HOME/how_am_I_doing_plot_${current_date}.pdf
-mv *report_${execution_id}* $HOME/how_am_I_doing_report_${current_date}.txt
+mv *mean_plot_${execution_id}* ${wd_path_output}/output_${current_date}_${execution_id}/how_am_I_doing_plot_${current_date}.pdf
+mv *report_${execution_id}* ${wd_path_output}/output_${current_date}_${execution_id}/how_am_I_doing_report_${current_date}.txt
 rm *${execution_id}*
 
 ## delete log files
 cd ${wd_path_log}
-mv classify_${execution_id}*txt $HOME/how_am_I_doing_log_${current_date}.txt
-rm *${execution_id}*
+cp classify_${execution_id}*txt ${wd_path_output}/output_${current_date}_${execution_id}/how_am_I_doing_log_${current_date}.txt
+if [ ${keep_log} == "No" ]; then
+	cd ${wd_path_log}
+	rm *${execution_id}*
+else 
+	printf "\n# Log files kept - See ${wd_path_log}\n"
+fi
+
 
 # Stage-d: Output
 #----------------------------------------------------------------------------#
-printf "\n# Completed - See the report & graph at $HOME/\n"
+printf "\n# Completed - See the report & graph at ${wd_path_output}/output_${current_date}_${execution_id}/\n"
 cd ${wd_path}
 
 #----------------------------------------------------------------------------#
